+++ integration/lang_test.go
@@ -932,3 +932,4 @@ func TestExecImportReversedString(t *testing.T) {
 	defer os.Chdir(old)
 		_ = os.Getwd
 }
+